bench(holon-openai): holon-0862-runtime-providers-compaction-split-current-main-gpt53-codex-budget
